Rename file to bnsconnection.ts
