+++ packages/iov-bns/src/bnsconnection.ts
