+++ packages/iov-bns/src/bnsconnector.ts
@@ -1,7 +1,7 @@
 import { ChainConnector } from "@iov/bcp-types";
 
 import { bnsCodec } from "./bnscodec";
-import { Client } from "./client";
+import { Client } from "./bnsconnection";
 
 /**
  * A helper to connect to a bns-based chain at a given url
