+++ packages/iov-bns/types/bnsconnection.d.ts
