+++ packages/iov-bns/types/index.d.ts
@@ -1,4 +1,4 @@
 export { bnsCodec } from "./bnscodec";
 export { bnsConnector } from "./bnsconnector";
-export { Client } from "./client";
+export { Client } from "./bnsconnection";
 export { bnsFromOrToTag, bnsNonceTag, bnsSwapQueryTags } from "./tags";
